data_engineering/airflow_project/requirements.txt
pyspark==3.2.1
apache-airflow
pyarrow==8.0.0
boto3
requests